requirements.txt
flask
requests
PyGithub
python-dotenv
